Pass ufbxw_string to ufbxwi_dom_value by value in save_mesh
Fixes segfault on Mac / invalid write call on Linux when saving meshes (Linux seems to not crash when invalid pointers and lengths are passed to `write()`)

attrib->name was being passed by pointer. You can see that the va_arg call uses the struct type itself, not the pointer, so this must also pass in the struct itself.

I checked and this is the only place matching the regex `ufbxwi_dom_value.*&` so probably there are no other occurrences of this issue.